// src/shared/api/models/postApiProductCategories201.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiProductCategories201Description } from './postApiProductCategories201Description';

export type PostApiProductCategories201 = {
    id: string;
    /**
     * @minimum -2147483648
     * @maximum 2147483647
     */
    seqId: number;
    name: string;
    description: PostApiProductCategories201Description;
    createdAt: string;
    updatedAt: string;
};